Struct poli_math::Euler

pub struct Euler {
    pub x: f32,
    pub y: f32,
    pub z: f32,
    pub order: EulerOrder,
}

Euler angles.

Euler angles describe a rotation in 3D space by the amount of rotation by each local axis and an axis order. They are intuitive to use but suffer from the gimbal lock problem.

For a better representation of rotations, use Quaternion, which represents a rotation around an arbitrary axis.

You can create Euler angles from a 4-tuple, a 3-tuple, or an array.

let a: Euler = (1, 2, 3, "XYZ").into();
let b: Euler = (4, 5, 6).into(); // uses default angle order XYZ
let c: Euler = [7, 8, 9].into(); // uses default angle order XYZ

Fields
x: f32

Angle of the X axis in radians.

y: f32

Angle of the Y axis in radians.

z: f32

Angle of the Z axis in radians.

order: EulerOrder

Order that the rotations are applied.

Implementations

impl Euler

pub fn from_rotation_matrix(m: &Matrix4, order: EulerOrder) -> Self

Creates Euler angles from the given rotation matrix and axis order.

pub fn from_quaternion(q: &Quaternion, order: EulerOrder) -> Self

Creates Euler angles from the given rotation quaternion and axis order.

pub fn set(&mut self, x: f32, y: f32, z: f32, order: Option<EulerOrder>)

Sets the X, Y, and Z angles, and optionally the order.