Struct poli_math::Matrix4

#[repr(C)]
pub struct Matrix4 {
    pub elements: [f32; 16],
}

4x4 matrix, commonly used to encode transformations i.e. translation, rotation, and scale. Note that elements are stored in column-major order due to the fact that WGSL uses column-major ordering.

Fields
elements: [f32; 16]

The elements of this matrix in column-major order.

Implementations

impl Matrix4

pub fn new( n11: f32, n12: f32, n13: f32, n14: f32, n21: f32, n22: f32, n23: f32, n24: f32, n31: f32, n32: f32, n33: f32, n34: f32, n41: f32, n42: f32, n43: f32, n44: f32, ) -> Self

Creates a new 4x4 matrix with the given row-major elements. The elements will be stored internally in column-major order.

pub fn from_translation(v: &Vector3) -> Self

Returns the translation matrix of the given displacement vector.

pub fn from_rotation_x(theta: f32) -> Self

Returns the rotation matrix around the X axis by the given angle in radians.

pub fn from_rotation_y(theta: f32) -> Self

Returns the rotation matrix around the Y axis by the given angle in radians.

pub fn from_rotation_z(theta: f32) -> Self

Returns the rotation matrix around the Z axis by the given angle in radians.

pub fn from_euler(euler: &Euler) -> Self

Returns the rotation matrix from the given Euler angles.

The implementation is based on formulae on this page.

pub fn from_quaternion(q: &Quaternion) -> Self

Returns the rotation matrix from the given rotation quaternion.

The implementation is based on the formulae on this page.

pub fn from_scale(v: &Vector3) -> Self

Returns the transformation matrix for the given scale transform.

pub fn compose( translation: &Vector3, rotation: &Quaternion, scale: &Vector3, ) -> Self

Creates a matrix for the transformation composed of the given translation, rotation, and scale. This uses TRS ordering: scale first, then rotation, then translation.

The implementation is based on the formulae for quaternion to 4x4 matrix conversion, with the addition of scale and translation.

pub fn look_at(eye: &Vector3, target: &Vector3, up: &Vector3) -> Self

Returns a rotation matrix looking from eye towards target oriented by the up vector.

pub fn set( &mut self, n11: f32, n12: f32, n13: f32, n14: f32, n21: f32, n22: f32, n23: f32, n24: f32, n31: f32, n32: f32, n33: f32, n34: f32, n41: f32, n42: f32, n43: f32, n44: f32, ) -> &Self

Sets the elements of this matrix with the given row-major elements.

pub fn copy(&mut self, m: &Self) -> &Self

Copies the elements from another matrix into this matrix.

pub fn identity() -> Self

Returns the 4x4 identity matrix.

pub fn zero() -> Self

Returns the 4x4 zero matrix.

pub fn translation(&self) -> Vector3

Returns the translation component of this matrix.

pub fn translate(&mut self, v: &Vector3) -> &Self

Translates by the given vector.

pub fn rotate(&mut self, q: &Quaternion) -> &Self

Applies the given rotation quaternion.

If you have Euler angles, you can use [Euler::into].

pub fn scale(&mut self, v: &Vector3) -> &Self

Scales by the given scale factor.

pub fn premultiply(&mut self, m: &Self) -> &Self

Pre-multiplies this matrix with the given matrix.

pub fn determinant(&self) -> f32

Returns the determinant of this matrix. Note that this does not modify this matrix.

The algorithm can be found here.

pub fn adjugate(&self) -> Self

Returns the adjugate of this matrix.

pub fn inverse(&self) -> Self

Returns the inverse of this matrix. If this matrix has no inverse i.e. the determinant is zero, then return the 4x4 zero matrix.

The inverse is calculated in terms of its adjugate.